# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJN Luxury Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP via WhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current customer
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── PRODUCTS ───────────────────────────────────────
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── ORDERS ──────────────────────────────────────────
...
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Create gallery item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete gallery item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Customer token returned by /auth/verify-otp.
    AdminCookieAuth:
      type: apiKey
      in: cookie
      name: ajn_admin_session
      description: HttpOnly admin session cookie returned by /admin/auth/login.
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    OtpResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        devOtp:
          type: ["string", "null"]
    OtpVerify:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    AuthResult:
      type: object
      required: [customer, token]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        token:
          type: string
    Customer:
      type: object
      required: [id, phone, name, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string

    # ─── PRODUCTS ─────────────────────
    Product:
      type: object
      required: [id, name, nameAr, price, stock, images, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        stock:
          type: integer
        category:
          type: ["string", "null"]
        subcategory:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
...
    Cart:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
    CartItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
        selectedColor:
          type: ["string", "null"]
        customization:
          type: ["string", "null"]
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        selectedColor:
          type: string
        customization:
          type: string
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    # ─── ORDERS ───────────────────────
...
    GalleryItem:
      type: object
      required: [id, mediaUrl, mediaType, category, createdAt]
      properties:
        id:
          type: integer
        mediaUrl:
          type: string
        mediaType:
          type: string
        title:
          type: ["string", "null"]
        titleAr:
          type: ["string", "null"]
        category:
          type: string
        createdAt:
          type: string
    GalleryItemInput:
      type: object
      required: [mediaUrl, mediaType, category]
      properties:
        mediaUrl:
          type: string
        mediaType:
          type: string
        title:
          type: string
        titleAr:
          type: string
        category:
          type: string

    # ─── REVIEWS ──────────────────────
...