# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AJN Luxury Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";

export interface AuthResult {
  customer: Customer;
  token: string;
}